[#1] style: move login button to introduction screen
